config and setting classes moved to dedicated files
