+++ src/NServiceBus.Core/Recoverability/Settings/DelayedRetriesSettings.cs
@@ -1,0 +1,40 @@
+namespace NServiceBus
+{
+    using System;
+    using Configuration.AdvanceExtensibility;
+    using Settings;
+
+    /// <summary>
+    /// Configuration settings for delayed retries.
+    /// </summary>
+    public class DelayedRetriesSettings : ExposeSettings
+    {
+        internal DelayedRetriesSettings(SettingsHolder settings) : base(settings)
+        {
+        }
+
+        /// <summary>
+        /// Configures the number of times a message should be retried with a delay after failing first level retries.
+        /// </summary>
+        public DelayedRetriesSettings NumberOfRetries(int numberOfRetries)
+        {
+            Guard.AgainstNegative(nameof(numberOfRetries), numberOfRetries);
+
+            Settings.Set(Recoverability.SlrNumberOfRetries, numberOfRetries);
+
+            return this;
+        }
+
+        /// <summary>
+        /// Configures the delay interval increase for each failed second level retry attempt.
+        /// </summary>
+        public DelayedRetriesSettings TimeIncrease(TimeSpan timeIncrease)
+        {
+            Guard.AgainstNegative(nameof(timeIncrease), timeIncrease);
+
+            Settings.Set(Recoverability.SlrTimeIncrease, timeIncrease);
+
+            return this;
+        }
+    }
+}
