+++ src/NServiceBus.Core/Recoverability/Settings/ImmediateRetriesSettings.cs
@@ -1,0 +1,27 @@
+namespace NServiceBus
+{
+    using Configuration.AdvanceExtensibility;
+    using Settings;
+
+    /// <summary>
+    /// Configuration settings for immediate retries.
+    /// </summary>
+    public class ImmediateRetriesSettings : ExposeSettings
+    {
+        internal ImmediateRetriesSettings(SettingsHolder settings) : base(settings)
+        {
+        }
+
+        /// <summary>
+        /// Configures the amount of times a message should be immediately retried after failing before escalating to second level
+        /// retries.
+        /// </summary>
+        /// <param name="numberOfRetries">The number of times to immediately retry a failed message.</param>
+        public void NumberOfRetries(int numberOfRetries)
+        {
+            Guard.AgainstNegative(nameof(numberOfRetries), numberOfRetries);
+
+            Settings.Set(Recoverability.FlrNumberOfRetries, numberOfRetries);
+        }
+    }
+}
