+++ src/NServiceBus.Core/Recoverability/Settings/RetryFailedSettings.cs
@@ -1,0 +1,30 @@
+namespace NServiceBus
+{
+    using System;
+    using System.Collections.Generic;
+    using Configuration.AdvanceExtensibility;
+    using Settings;
+
+    /// <summary>
+    /// Configuration settings for retry faults.
+    /// </summary>
+    public class RetryFailedSettings : ExposeSettings
+    {
+        internal RetryFailedSettings(SettingsHolder settings) : base(settings)
+        {
+        }
+
+        /// <summary>
+        /// Configures a header customization action which gets called after all fault headers have been applied.
+        /// </summary>
+        /// <param name="customization">The customization action.</param>
+        public RetryFailedSettings HeaderCustomization(Action<Dictionary<string, string>> customization)
+        {
+            Guard.AgainstNull(nameof(customization), customization);
+
+            Settings.Set(Recoverability.FaultHeaderCustomization, customization);
+
+            return this;
+        }
+    }
+}
